+++ doc/Changelog.md
@@ -3,6 +3,9 @@ Changelog {#Changelog}
 
 # git master {master)
 
+* [6](https://github.com/BlueBrain/Monsteer/pull/6):
+  Add monsteerGUI application + widget to receive the selected cell ids and
+  inject stimuli into a NEST simulation. Available as a library (MonsteerQt)
 * [15819](http://bbpcode.epfl.ch/code/#/c/15819)
   Add Nestifier class to convert BBP to NEST circuits
 * [#4](https://github.com/BlueBrain/Monsteer/pull/4)